Fix deadlock when disposing internal consumers
