+++ Source/EasyNetQ.Tests/ModelCleanupTests.cs
@@ -9,81 +9,16 @@ namespace EasyNetQ.Tests
 {
     public class ModelCleanupTests
     {
-        private readonly IBus bus;
-        private readonly MockBuilder mockBuilder;
-        private readonly TimeSpan waitTime;
-
         public ModelCleanupTests()
         {
             mockBuilder = new MockBuilder();
             bus = mockBuilder.Bus;
-            waitTime = TimeSpan.FromMinutes(2);
+            waitTime = TimeSpan.FromSeconds(10);
         }
 
-        [Fact]
-        public void Should_not_cleanup_publish_model()
-        {
-            bus.Publish(new TestMessage());
-            bus.Dispose();
-
-            mockBuilder.Channels[0].DidNotReceive().Dispose();
-        }
-
-        [Fact]
-        public void Should_cleanup_subscribe_model()
-        {
-            bus.Subscribe<TestMessage>("abc", mgs => {});
-            var are = WaitForConsumerModelDisposedMessage();
-
-            bus.Dispose();
-
-            bool signalReceived = are.WaitOne(waitTime);
-            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
-
-            mockBuilder.Channels[1].Received().Dispose();
-        }
-
-        [Fact]
-        public void Should_cleanup_subscribe_async_model()
-        {
-            bus.SubscribeAsync<TestMessage>("abc", msg => null);
-            var are = WaitForConsumerModelDisposedMessage();
-
-            bus.Dispose();
-
-            bool signalReceived = are.WaitOne(waitTime);
-            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
-
-            mockBuilder.Channels[1].Received().Dispose();
-        }
-
-        [Fact]
-        public void Should_cleanup_request_response_model()
-        {
-            bus.RequestAsync<TestRequestMessage, TestResponseMessage>(new TestRequestMessage());
-            var are = WaitForConsumerModelDisposedMessage();
-
-            bus.Dispose();
-
-            bool signalReceived = are.WaitOne(waitTime);
-            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
-
-            mockBuilder.Channels[1].Received().Dispose();
-        }
-
-        [Fact]
-        public void Should_cleanup_respond_model()
-        {
-            bus.Respond<TestRequestMessage, TestResponseMessage>(x => null);
-            var are = WaitForConsumerModelDisposedMessage();
-
-            bus.Dispose();
-
-            bool signalReceived = are.WaitOne(waitTime);
-            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
-
-            mockBuilder.Channels[1].Received().Dispose();
-        }
+        private readonly IBus bus;
+        private readonly MockBuilder mockBuilder;
+        private readonly TimeSpan waitTime;
 
         private AutoResetEvent WaitForConsumerModelDisposedMessage()
         {
@@ -92,6 +27,83 @@ namespace EasyNetQ.Tests
             mockBuilder.EventBus.Subscribe<ConsumerModelDisposedEvent>(x => are.Set());
 
             return are;
+        }
+
+        [Fact]
+        public void Should_cleanup_publish_model()
+        {
+            bus.Publish(new TestMessage());
+            bus.Dispose();
+
+            mockBuilder.Channels[0].Received().Dispose();
+        }
+
+        [Fact]
+        public void Should_cleanup_request_response_model()
+        {
+            var waiter = new CountdownEvent(2);
+
+            mockBuilder.EventBus.Subscribe<PublishedMessageEvent>(_ => waiter.Signal());
+            mockBuilder.EventBus.Subscribe<StartConsumingSucceededEvent>(_ => waiter.Signal());
+
+            bus.RequestAsync<TestRequestMessage, TestResponseMessage>(new TestRequestMessage());
+            if (!waiter.Wait(5000))
+                throw new TimeoutException();
+
+            var are = WaitForConsumerModelDisposedMessage();
+
+            bus.Dispose();
+
+            var signalReceived = are.WaitOne(waitTime);
+            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
+
+            mockBuilder.Channels[0].Received().Dispose();
+            mockBuilder.Channels[1].Received().Dispose();
+        }
+
+        [Fact]
+        public void Should_cleanup_respond_model()
+        {
+            bus.Respond<TestRequestMessage, TestResponseMessage>(x => (TestResponseMessage) null);
+            var are = WaitForConsumerModelDisposedMessage();
+
+            bus.Dispose();
+
+            var signalReceived = are.WaitOne(waitTime);
+            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
+
+            mockBuilder.Channels[0].Received().Dispose();
+            mockBuilder.Channels[1].Received().Dispose();
+        }
+
+        [Fact]
+        public void Should_cleanup_subscribe_async_model()
+        {
+            bus.Subscribe<TestMessage>("abc", msg => { });
+            var are = WaitForConsumerModelDisposedMessage();
+
+            bus.Dispose();
+
+            var signalReceived = are.WaitOne(waitTime);
+            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
+
+            mockBuilder.Channels[0].Received().Dispose();
+            mockBuilder.Channels[1].Received().Dispose();
+        }
+
+        [Fact]
+        public void Should_cleanup_subscribe_model()
+        {
+            bus.Subscribe<TestMessage>("abc", mgs => { });
+            var are = WaitForConsumerModelDisposedMessage();
+
+            bus.Dispose();
+
+            var signalReceived = are.WaitOne(waitTime);
+            Assert.True(signalReceived, $"Set event was not received within {waitTime.TotalSeconds} seconds");
+
+            mockBuilder.Channels[0].Received().Dispose();
+            mockBuilder.Channels[1].Received().Dispose();
         }
     }
 }
