+++ Source/EasyNetQ/Consumer/IConsumerDispatcher.cs
@@ -5,6 +5,7 @@ namespace EasyNetQ.Consumer
     public interface IConsumerDispatcher : IDisposable
     {
         void QueueAction(Action action);
+        void QueueAction(Action action, bool surviveDisconnect);
         void OnDisconnected();
     }
 }
